Carry over computer id when regenerating sessions (#57)
This is crucial because the computer id is saved to the session
asynchronously and when the session is regenerated, for the following
queries it would be null. This fixes people being able to withdraw
each other's questions. This was mostly affecting logged in users
or users who have logged in at least once.